HZN-1170: Some cleanup
