--- features/flows/api/src/main/java/org/opennms/netmgt/flows/api/AbstractFlowDocument.java
@@ -1,70 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2017-2017 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2017 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.netmgt.flows.api;
-
-import com.google.gson.annotations.SerializedName;
-
-public abstract class AbstractFlowDocument implements FlowDocument {
-
-    @SerializedName("location")
-    private String location;
-
-    @SerializedName("exporter_address")
-    private String exporterAddress;
-
-    @SerializedName("flow_type")
-    private String flowType;
-
-    @Override
-    public String getLocation() {
-        return this.location;
-    }
-
-    @Override
-    public String getExporterAddress() {
-        return this.exporterAddress;
-    }
-
-    @Override
-    public String getFlowType() {
-        return flowType;
-    }
-
-    public void setLocation(String location) {
-        this.location = location;
-    }
-
-    public void setExporterAddress(String exporterAddress) {
-        this.exporterAddress = exporterAddress;
-    }
-
-    public void setFlowType(String flowType) {
-        this.flowType = flowType;
-    }
-}
